oprava modelu

+++ src/Model/AclResources/AclResourcesMapper.php
@@ -35,7 +35,9 @@ class AclResourcesMapper extends Mapper
 			/* @var $rule Acl */
 			$resources[] = $rule->resource->id;
 		}
-		$this->connection->query('DELETE FROM %table WHERE [id] NOT IN %i[]', $this->getTableName(), $resources);
+		if (!empty($resources)) {
+			$this->connection->query('DELETE FROM %table WHERE [id] NOT IN %i[]', $this->getTableName(), $resources);
+		}
 	}
 
 }
